login and registration moved to modals on welcome page - users create default account on registration - can now have empty set of transactions
