+++ src/main/java/com/bankingapp/backend/controller/RegisterController.java
@@ -18,53 +18,5 @@ import java.util.List;
 
 @Controller
 public class RegisterController {
-    @Autowired
-    private NewAccountService newacc;
-
-    @Autowired
-    private NewUserService newuser;
-    /* bandaid*/
-    private int countryId = 0;
-
-    @GetMapping("/register")
-    public String showRegistration(Model model) {
-        Customer customer = new Customer();
-        model.addAttribute("customer", customer);
-
-        List<String> listCountyState = Arrays.asList("Dublin, Ireland", "Galway, Ireland", "Cork, Ireland");
-        model.addAttribute("listCountyState", listCountyState);
-
-        return "registration";
-    }
-
-    @PostMapping("/register")
-    public String addCustomer(
-            @RequestParam String firstName,
-            @RequestParam String lastName,
-            @RequestParam String email,
-            @RequestParam int phoneNumber,
-            @RequestParam String dob,
-            @RequestParam String addrLine1,
-            @RequestParam String addrLine2,
-            @RequestParam String townCity,
-            @RequestParam String countyState,
-            @RequestParam String password,
-            @RequestParam String idType,
-            @RequestParam String idNumber) {
-
-
-        Customer newCustomer = new Customer(firstName,lastName, email, phoneNumber, countryId, addrLine1, addrLine2, townCity, countyState, password, idType, idNumber, dob);
-        /* this should return failure  if it was unsucessful to create new user */
-        newuser.addNewCustomer(newCustomer);
-
-        return "welcome";
-    }
-
-    @GetMapping("/debug-list")
-    public String getAllCustomers(Model model) {
-        List<Customer> customers = newuser.getAllCustomers();
-        model.addAttribute("customers", customers);
-        return "debug-list";
-    }
 
 }
