+++ src/main/java/com/bankingapp/backend/service/NewAccountService.java
@@ -7,5 +7,7 @@ import java.util.List;
 public interface NewAccountService {
 
     List<Account> getTransactions(long id);
+    int getCustomerIdForAccount();
+    Account addNewAccount(Account account);
 
 }
