+++ src/main/java/com/bankingapp/backend/service/NewAccountServiceImpl.java
@@ -1,7 +1,9 @@
 package com.bankingapp.backend.service;
 
 import com.bankingapp.backend.model.Account;
+import com.bankingapp.backend.model.Customer;
 import com.bankingapp.backend.repository.AccountRepository;
+import com.bankingapp.backend.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
@@ -15,7 +17,11 @@ public class NewAccountServiceImpl  implements NewAccountService {
     private final AccountRepository accountRepository;
 
     @Autowired
-    public NewAccountServiceImpl( AccountRepository accountRepository) {
+    private final CustomerRepository customerRepository;
+
+    @Autowired
+    public NewAccountServiceImpl( AccountRepository accountRepository, CustomerRepository customerRepository) {
+        this.customerRepository = customerRepository;
         this.accountRepository = accountRepository;
     }
 
@@ -25,4 +31,24 @@ public class NewAccountServiceImpl  implements NewAccountService {
         Specification<Account> isTransaction = AccountSpecifications.accountTransactions(id);
         return accountRepository.findAll(isTransaction);
     }
+
+    @Override
+    public int getCustomerIdForAccount() {
+        int index = -1;
+        List<Customer> lastCustomer = customerRepository.findAll();
+
+        for (int i = 0; i <lastCustomer.size(); i++){
+            index++;
+            System.out.println(index);
+        }
+
+        return lastCustomer.get(index).getCustomerId();
+    }
+
+    @Override
+    public Account addNewAccount(Account account) {
+        return accountRepository.save(account);
+    }
+
+
 }
