+++ src/main/resources/templates/welcome.html
@@ -3,23 +3,20 @@
 <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz" crossorigin="anonymous"></script>
+    <!-- <script src="../scripts/script.js"></script>
+
+    <link rel="stylesheet" th:href="@{/css/main.css}"/> -->
 
     <title>An amazing Bank Application</title>
-
-    <link rel="stylesheet" th:href="@{webjars/bootstrap/4.2.1/css/bootstrap.min.css}"/>
-    <link rel="stylesheet" th:href="@{/css/main.css}"/>
-
 </head>
-
 <body>
-
 <nav class="navbar navbar-expand-md navbar-dark bg-dark fixed-top">
     <a class="navbar-brand" href="#">Project Bank</a>
     <a class="navbar-brand" href="/debug-list">Current customers debug page</a>
 </nav>
-
 <main role="main" class="container">
-
     <div style="margin-left:auto;margin-right:auto;text-align:center" class="starter-template">
         <h1>An amazing Bank Application</h1>
     </div>
@@ -29,14 +26,95 @@
     </ol>
      -->
     <div align="center">
-        <h1>Registration Form</h1>
-        <h2><a href="/login">Log In</a></h2>
-        <a href="/register">Click here to register</a>
+        <h1>Welcome to The Bank of Galway</h1>
+        <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#loginModal">Login</button><br><br>
+        <button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#regModal">Register</button><br>
     </div>
+    <!-- Button trigger modal -->
 
+
+    <!-- Login Modal -->
+    <div class="modal fade" id="loginModal" tabindex="-1" aria-labelledby="loginModalLabel" aria-hidden="true">
+        <div class="modal-dialog">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <h1 class="modal-title fs-5" id="loginModalLabel">Login</h1>
+                    <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                </div>
+                <div class="modal-body">
+                    <form action="#" th:action="@{/dashboard/}">
+                        <p><input type="text" th:name="CustomerId"></p>
+                        <p><input type="submit" value="Submit" /></p>
+                    </form>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
+                </div>
+            </div>
+        </div>
+    </div>
+</main>
+
+<!-- Registration Modal -->
+<div class="modal fade" id="regModal" tabindex="-1" aria-labelledby="regModalLabel" aria-hidden="true">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h1 class="modal-title fs-5" id="regModalLabel">Register</h1>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+            <div class="modal-body">
+                <form action="#" th:action="@{/}" method="post" th:object="${customer}">
+                    <label>First name:</label>
+                    <input type="text" th:field="*{firstName}"/><br/>
+
+                    <label>Last name:</label>
+                    <input type="text" th:field="*{lastName}"/><br/>
+
+                    <label>E-mail:</label>
+                    <input type="text" th:field="*{email}"/><br/>
+
+                    <label>Password:</label>
+                    <input type="text" th:field="*{password}"/><br/>
+
+                    <label>Phone Number:</label>
+                    <input type="text" th:field="*{phoneNumber}"/><br/>
+
+                    <label>Address Line 1</label>
+                    <input type="text" th:field="*{addrLine1}"/><br/>
+
+                    <label>Address Line 2</label>
+                    <input type="text" th:field="*{addrLine2}"/><br/>
+
+                    <label>Town/City</label>
+                    <input type="text" th:field="*{townCity}"/><br/>
+
+                    <label>County/State:</label>
+                    <select th:field="*{countyState}">
+                        <option th:each="p : ${listCountyState}" th:value="${p}" th:text="${p}"/>
+                    </select><br/>
+
+                    <label>id Type</label>
+                    <input type="text" th:field="*{idType}"/><br/>
+
+                    <label>ID Number</label>
+                    <input type="text" th:field="*{idNumber}"/><br/>
+
+                    <label>DOB</label>
+                    <input type="text" th:field="*{dob}"/><br/>
+
+                    <button type="submit">Register</button>
+                </form>
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                <button type="button" class="btn btn-primary">Save changes</button>
+            </div>
+        </div>
+    </div>
+</div>
 </main>
 <!-- /.container -->
-
 <script type="text/javascript" th:src="@{webjars/bootstrap/4.2.1/js/bootstrap.min.js}"></script>
 </body>
 </html>
